[svn r1227] fixes for ie, need to move to native vml implementation

+++ Makefile
@@ -2,7 +2,7 @@
 DIST_DIR=myWebDevelopment-`date +%Y%m%d`
 dist:
 	ln -s ./ myWebDevelopment
-	rm $(DIST_DIR).zip
+	rm -f $(DIST_DIR).zip
 	svn -R list | sed s/^/myWebDevelopment\\\//g | zip -@ $(DIST_DIR).zip
-	rm myWebDevelopment
+	if test -L myWebDevelopment; then rm myWebDevelopment; fi
